Completed Ep 30

+++ 5.7FromScratch/project/app/Http/Controllers/ProjectsController.php
@@ -4,13 +4,21 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Project;
+use App\Mail\ProjectCreated;
 
 class ProjectsController extends Controller
 {
     //
+
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function index() {
-        $projects = Project::all();
-        return view('projects.index', ['projects' => $projects]);
+        return view('projects.index', [
+            'projects' => auth()->user()->projects
+        ]);
     }
 
     public function create() {
@@ -18,31 +26,47 @@ class ProjectsController extends Controller
     }
 
     public function show(Project $project) {
+        $this->authorize('update', $project);
         return view('projects.show', compact('project'));
     }
 
     public function edit(Project $project) {
+        $this->authorize('update', $project);
         return view('projects.edit', compact('project'));
     }
 
     public function update(Project $project) {
-        $project->update(request(['title', 'description']));
-        return redirect('/projects');
-    }
-
-    public function destroy(Project $project) {
-        $project->delete();
-        return redirect('/projects');
-    }
-
-    public function store() {
+        $this->authorize('update', $project);
 
         $attributes = request()->validate([
             'title' => ['required', 'min:3', 'max:255'],
             'description' => ['required', 'min:3', 'max:255']
         ]);
 
-        Project::create($attributes);
+        $project->update($attributes);
+        return redirect('/projects');
+    }
+
+    public function destroy(Project $project) {
+        $this->authorize('update', $project);
+        $project->delete();
+        return redirect('/projects');
+    }
+
+    public function store() {
+        $attributes = request()->validate([
+            'title' => ['required', 'min:3', 'max:255'],
+            'description' => ['required', 'min:3', 'max:255']
+        ]);
+
+        $attributes['owner_id'] = auth()->id();
+
+        $project = Project::create($attributes);
+
+        \Mail::to($project->owner->email)->send(
+            new ProjectCreated($project)
+        );
+
         return redirect('/projects');
     }
 }
